Adding storage for the chimeric ledgers

+++ cef/network/src/main/scala/io/iohk/cef/ledger/chimeric/package.scala
@@ -3,7 +3,7 @@ package io.iohk.cef.ledger
 package object chimeric {
 
   type Currency = String
-  type Quantity = Int
+  type Quantity = BigDecimal
   type ChimericTxId = String
   type Address = String
 }
